backend/requirements-updated.txt
# Web Framework & Server
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0

# LLM & AI
anthropic>=0.25.0
openai>=1.0.0

# Configuration & Environment
python-dotenv>=1.0.0

# Computer Vision
opencv-python>=4.9.0
numpy>=1.26.0

# MQTT Communication
paho-mqtt>=1.6.0

# Utilities
pydantic>=2.0.0
requests>=2.31.0